Interesting results in tut-13 and neutronics-4-group example.
AMESOS gives different results in the complex tutorial than the other solvers.
AztecOO with BiCGStab introduces spurious oscillations to the solution in the neutronics example.
Also included are preliminary timing results of the various solvers (noteworthy here is the very long assembly time of PETSc comparing to the other solvers).