Use the stacktrace functionality
This patch integrates the stacktrace functionality into the CMake buildsystem.
This means that it checks for binutils and if it's available on the system, it
enables its usage as well as linking with the right libraries. If it's not
available, then it will not be used, so that hermes still compiles on all
platforms (like Mac or MSVC).