Reenabling the H3D_COMPLEX compiler definition in hermes3d/CMake.common, adding H3D_REAL (required for hermes_common functions).
The previously used 'add_definitions' added -DH3D_COMPLEX to compile options for all files in all subdirectories of hermes3d, including 'src'. This caused the multiple definitions warning.
Now, it is added for each project separately, but not for the library itself (where CMake.common is not included and where the definition is added from CMakeLists.txt).