Update hermes2d, 3d and common from step-1
The previous commit resolved all conflicts by taking the files from the step-1
branch. This commit deletes hermes2d, hermes3d and hermes_common, and then
reintroduces them from the step-1 branch.

This means, that things that were only done in master (but not step-1) are
lost. Later on, we need to go over parts of this commit and reintroduce/update
those things.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>