Dealing with Teuchos_stacktrace when some necessary functions are missing.
-one issue : in hermes3d/CMakeLists.txt it is said that
"The function vasprintf() that is declared in stdio.h in GNU C library is missing elsewhere"

which I am not sure applies to Macs (as I do not know much about them), what I read on the
internet was that the functino vasprintf() originally was only in GNU C lib, so I am not sure.
Can somebody with a Mac check this? Thank you.