Stabilized-advection-reaction: Updating the exact solver.
Nodal points of one of the quadrature rules available in Mathematica, together with corresponding exact solution values and quadrature weights, may now be exported.
Also in this commit is a file with quadrature info for the Gauss-Kronrod rule with 50 Gauss points (for which the Kronrod extension yields a total of 10201 evaluation points).