Further improvement to rk_time_step().
The new time-stepping is implemented so far in the
examples butcher and 09-timedep. A comparison
shows that rk_time_step() with Butcher's table
corresponding to SDIRK-2 in example butcher is almost
twice slower than its (less general) implementation
in example butcher-comparison. So more work on
the rk_time_step() function is needed. A list
of things that need to be done is at the beginning
of the file hermes2d/src/runge_kutta.cpp.