Make Cython use correct integer division
In Python it is enough to do "from __future__ import division" at the top of
the file and then even evaluation will use it. In Cython however, this doesn't
work that way, and it needs to be injected into the actual code to be
evaluated. After this patch, the code works identically both in Python and
Cython. And it works correctly.