Make tutorial tests 50 and 52 more robust
First of all, one needs to use std::abs, instead of abs. This was caused by
removing "using namespace std" from the top of the file. Such usage of
namespaces is very very strongly discouraged by the Trilinos guidelines, due to
bugs like this one.

Another problem is that the eigenvectors are defined up to a sign, so this fact
was also incorporated into the tests.