A major rework of the neutronics weak forms.
My original intention was to have local data in each neutronics form that could be accessed by the integer e->elem_marker inside the form evaluation methods (thus saving time to convert the string marker to the integer one every time the form is evaluated).
This rework was triggered by my realization that element_markers_conversion cannot actually be accessed from a WeakForm descendant before constructing the DiscreteProblem. Therefore, in order to achieve the above goal, the user-marker_to_internal-marker conversion cannot be done at the forms' construction time as before.
So my second attempt  was to do the conversion during the first evaluation of the form. However, this was unsuccessful as well, because the evaluation methods must be 'const' and hence cannot change the state of the form.
After several other thought attempts (quite unclean), I finally gave up - for problems with small number of materials, the performance hit may not be so large anyway...

This commit also adds the capability to cout the MaterialPropertyMaps container (originally for debugging purposes) and define a 'Sigma_s_nnz_structure' matrix. This latter addition reflects the fact that quite often, the scattering matrix contains only a few nonzeros (usually the first subdiagonal) for all materials. Since there is a Jacobian/residual pair added for each entry of the scattering matrix, this could save some time. Similar thing may also be done to the fission data (nu, Sigma_f, chi) to avoid some unneeded FissionYieldForms.

There is also a new Matrix/VectorFormSurf desendant for vacuum b.c. These b.c. are actually implemented as Newton conditions, for whose residual part I originally used the WeakFormsH1::DefaultResidualSurf class. The problem was that this class cannot be used for systems, since it implicitly uses 'u_ext[0]', no matter which place in the matrix weak formulation it represents (i.e. the 'i' parameter in its construction).