Introducing const-correctness & change in Runge-Kutta's constructor.
*issues: Spaces are not assigned global DOFs in DiscreteProblem, instead they are
passed as const pointers, i.e. the DOFs will not change. But the global enumeration is necessary, so here are the options:

a) create local Space instances with global enumeration
b) somehow (bad habit) make enough parameters mutable, so that one can actually call assign_dofs()
c) also very bad habit - go through assembling, somehow calculate the global DOFs out of the ordering in
   the passed vector of Spaces etc..
d) what else?
e) use const_cast to enumerate DOFs globally at the beginning of assemble(), and at the end call it again, to enumerate it back locally (or as it used to be before the call to assemble())

Lukas